[DotNet] Add signed narrow scalars
Map Kotlin Byte and Short to exact CLR int8 and int16 carriers while preserving Common promotion, conversion, nullable, boxing, generic, and overload semantics.

Normalize boxed hash and invariant string behavior across Framework CLR and CoreCLR, and accept matching foreign SByte and Int16 signatures. Freeze the ABI with adversarial compiler, IL, cross-module, and C# product tests on both profiles.

+++ compiler/testData/codegen/dotnet/ilText/nestedObjectDeclarationsRejected.kt
@@ -11,7 +11,7 @@ object BrokenObjectOwner {
     }
 
     object BrokenObject {
-        fun unsupported(value: Byte): Byte = value
+        fun unsupported(value: Float): Float = value
     }
 
     object GoodObject {
@@ -23,7 +23,7 @@ object BrokenObjectOwner {
     }
 
     open class BrokenBase {
-        fun unsupported(value: Byte): Byte = value
+        fun unsupported(value: Float): Float = value
     }
 
     class Derived : BrokenBase()
